requirements.txt
streamlit==1.28.0
google-auth==2.23.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
python-dateutil==2.8.2
pytz==2023.3
google-generativeai==0.3.2
groq==0.11.0
streamlit-modal==0.1.0
audio-recorder-streamlit
python-dotenv